Minor code analysis fixes

+++ src/RabbitMQ.Next.Transport/Methods/Channel/CloseMethodFormatter.cs
@@ -6,7 +6,7 @@ namespace RabbitMQ.Next.Transport.Methods.Channel
     internal class CloseMethodFormatter : IMethodFormatter<CloseMethod>
     {
         public Span<byte> Write(Span<byte> destination, CloseMethod method) =>
-            destination.Write((ushort) method.StatusCode)
+            destination.Write(method.StatusCode)
                 .Write(method.Description)
                 .Write((uint)method.FailedMethodId);
     }
